Create script.txt
select the curve you want to edit in object mode, go to edit mode and select one of the latest points as active, it will make curve go into one or another direction. Then exit edit mode and run the script with curve selected. It will apply the bendy bones with some given in the script parameters. Number of bones equal to the number of points minus one. With zero gravity it will keep form, with some gravitation used- it will need some bones with collision inside the object.